Change the sed and recall regexen to match the entire line.
Matcher.match() attempts to match the entire input line, effectively surrounding
the provided regex with '^' and '$' anchors. There's a find() method which
performs a match at the beginning of the input line, but this is stateful, as
repeated calls to find will iterate over all matches for a given regex in the
input line.

This means that the following forms don't work

  <cat> something interesting
  <dog> p/inter/cat  i agree

as the search regexes no longer match the entire line.

A different way to do this would be to change the appropriate matches()
invocations to call find() instead; they're probably more or less equivalent up
to the extra matching effort which matches() needs to perform for the trailing
'.*', given that the regexes in use are already caret-anchored.