Change partial line matches to use find().
The Matcher.matches() method attempts to match a regex over an entire string,
which results in a trailing '.*' for regexen which only need to partially match
the string. Matcher.find(), on the other hand, attempts to find the next match
in a given string, so it's probably marginally more efficient to use that rather
than process the '.*' necessary to make things work with a whole-string match.