Refactor regular expressions.
Switch to using Google's re2j regex library instead of the Java
standard library. The Java standard library regex implementation is
backtracking, which means that the worst case time to perform a match
is exponential in the length of the inputs. Given that the whole point
of timemachine is to run arbitrary regular expressions, a user could
send timemachine a pathological regex which causes it to hang while
matching. This is a DoS vector which has been accidentally hit at least
once. re2j is based on non-deterministic finite automata, which takes
linear time to perform matches, which is much safer.

However, re2j doesn't support lookaround assertions, because
(as I understand it) such extensions can't be implemented without
backtracking. This poses an issue because the regexes which I borrowed
from Puck for matching the s/// and p// forms use a negative lookahead
assertion to prevent unescaped backslashes being matched. (To make
it absolutely clear: the regex library is being changed to remove a
DoS vector when running other users' regexes, however this library is
incompatible with the regexes I was using to decide when to trigger in
the first place.)

    Aside: A lookaround assertion is an extension to the regex syntax
    (which I *think* originates from Perl) which allows one to specify
    whether a pattern should match if it does or does not appear before
    or after some other pattern. In this case, a negative lookahead
    says that the preceding pattern matches only when it is *not*
    followed by the given pattern. For example

        /foo(?!bar)/

    will match the string "foo" *unless* it is immediately followed by
    the string "bar". The regex I originally borrowed from Puck
    contains the following pattern (which matches the "search"
    component of an s/// form):

        /(?:\\\/|[^\/])*(?!\\)/

    What this means is "Match zero or more occurences of either a
    backslash/forward slash pair or a character which isn't a forward
    slash. But only if it's *not* followed by a single backslash".

As a result, I've rewritten the regexes used for matching sed lines and
recall lines. Instead of matching "anything except unescaped forward
slashes not followed by a backslash", I'm now matching on "anything
character which isn't an unescaped slash (of either direction), or a
backslash paired with another character".

I also discovered some subtle behaviour surrounding trailing excess
delimiters. A message of the form "s/foo/bar////" will have the
trailing slashes captured in the third group (the nick against which to
perform the search-and-replace operation) by Puck's original regular
expression, and I had to gently tweak my replacement regex to match
this behaviour.

A second subtle behaviour is what happens in messages like
"s/foo/bar/~" which are missing trailing digits. This will match
up to and including the third slash, but leave the tilde unmatched;
effectively, having a tilde with no trailing digits is the same as
omitting the tilde or specifying "~0". I'm not quite sure what the
correct thing to do here is, as I could change the /[0-9]+/ match to
/[0-9]*/ instead to ensure that the tilde is always captured and then
handle the missing digits in some later validation step, or I could
add /(?:\s|$)/ to the end of the regex to ensure that a match of any
of the preceding patterns must be followed by whitespace or the end of
line. As there's no immediately clear benefit one way or another, I'm
going to leave it as it is for now.