Add /g switch to search/replace for replacing all occurences instead of just the first.
People keep expecting timemachine to support a sed-like "global" trailing
modifier, and as it turns out this isn't too hard to implement. This will break
using the trailing rename target syntax on people with the nick "g", though I
don't forsee this being much of an issue in the grand scheme of things.