Substantially rework message history tracking logic.
The primary user-visible change in this commit is that messages which
trigger the p// or s/// commands are now recorded in the channel message
history. Previously, if a message which triggered a command had a trailing
comment, e.g.:

    <jane> s/jpg/png/  i prefer lossless images

then the message would not be recorded at all, which would prevent users from
invoking commands on the trailing comment (in this case, "i prefer lossless
images").

To handle this case, channel messages have been extended with the notion of a
non-matching prefix. Normal messages have an empty prefix (and hence the entire
message text is matchable), but messages which contain matching commands are
split into the prefix containing the command text and the trailing comment (and
only the trailing comment can be matched by later commands).

For example:

-   <jane> i'm going to the supermarket, brb
           ---------------------------------
                         |
                    message body

           message prefix
                 |
            ----------
-   <alice> p/meeting/  reminder for tomorrow's meeting times
                      ---------------------------------------
                                        |
                                  message body

              message prefix
                    |
           -------------------
-   <eric> john: s/coffee/tea/  ftfy :D
                              ---------
                                  |
                             message body

This addresses a common complaint about the inability to recall other earlier
recall or sed messages based on the trailing text in the message.

This commit also changes the organisation of the channel message log. Prior
to change, channel messages were grouped on a per-user basis, however this
commit removes this grouping, and all messages in a given channel are stored
in a single shared message log. This removes some indirection and simplifies
some lookup code (at the expense of needing to traverse the entire channel log
on every command, instead of only the target user's message log), with the
main benefits being improved bslsk05 compatibility, and simplifying the future
implementation of some extensions to the p// syntax.