Fix null pointer dereference in search and replace.
In search and replace messages which include the trailing slash (e.g.
"s/foo/bar/"), the `target` match is initialised, as the command regex captures
an empty string. However, if the trailing slash is omitted ("s/foo/bar"), then
`target` is set to null. This causes the condition which evaluates whether to
perform a global replacement to dereference a null pointer.